# [LCS] faster serialization of maps