--- a/common/lcs/Cargo.toml
+++ b/common/lcs/Cargo.toml
@@ -14,5 +14,10 @@ thiserror = "1.0"
 serde = { version = "1.0", features = ["derive"] }
 
 [dev-dependencies]
+criterion = "=0.3.1"
 proptest = "0.9"
 proptest-derive = "0.1.1"
+
+[[bench]]
+name = "lcs_bench"
+harness = false