## Motivation

Inserting N times in a BTreeMap to sort a list of entries does not feel quite right. Let's simply sort of vector instead.

I also compared stable sort v.s. unstable sort. Picking the stable sort because it delivers a better win (-20% time) in the case of `BTreeMap` -- which I expect is our main use case in practice.

## Test Plan

```
$ cargo bench -p libra-canonical-serialization
...
     Running target/release/deps/lcs_bench-5cda580006e59e2b
serialize btree map     time:   [624.22 us 628.37 us 632.84 us]                                
                        change: [-22.655% -21.284% -20.011%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 1 outliers among 100 measurements (1.00%)
  1 (1.00%) high mild

serialize hash map      time:   [708.39 us 715.67 us 724.93 us]                               
                        change: [-16.926% -15.328% -13.683%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 10 outliers among 100 measurements (10.00%)
  2 (2.00%) high mild
  8 (8.00%) high severe
```