@matbd 2020-03-26
fixed formatting, add `dedup_by` to match the previous semantics exactly